requirement.txt
Django==2.0.7
django-bootstrap3==10.0.1
django-filter==2.0.0
django-widget-tweaks==1.4.2
mysqlclient==1.3.13
pytz==2018.5
